requirements.txt
Flask
Flask-Cors
python-dotenv
supabase
apscheduler
pytz
groq
smtplib  # (part of Python stdlib, no need to install)
jsonlib  # (optional, since `json` is stdlib, remove if not used)
